[PACHA-22] Disallow recursive types in SQL table column types (#1056)
<!-- The PR description should answer 2 important questions: -->

### What

- Disallow recursive types in types of table columns

### How

- As we walk down the table type constructing the corresponding Arrow
types, we keep track of a set of struct type names that we've seen.
- If we see a type name we've seen before, we don't include the current
field:
  - If we're in a nested context, we cascade the field deletion up
  - If we're at a top-level table column, we remove that column.

With this approach, the fields which are removed never depend on the
path the traversal takes, so we never end up with a reference to a named
struct type which in reality fetches only a subset of the total fields.
A named type always refers to the same set of fields.

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

V3_GIT_ORIGIN_REV_ID: 270a71cd5b1d3700067abf7c771c473bbc33167e