Don't include .git folder and instead provide RELEASE_VERSION env var in `Dockerfile` (#1094)
<!-- The PR description should answer 2 important questions: -->

### What

The `Dockerfile` for `v3-engine` is provided for local development, and
is not what we use for production builds. Therefore there is no need to
version stamp it, so instead, provide `RELEASE_VERSION=dev` and don't
include the `.git` folder as this breaks when building from the
`graphql-engine` repo.

### How

Changes to root `Dockerfile`.

V3_GIT_ORIGIN_REV_ID: 6ed7ad43456c709e24b7768417436dcf6123adfa